test(task-migrate-clean-batch): 迁 gui_state_reconcile 3 用例到 gui_unit_test
420.5 Step 3。整文件迁移。该文件已是"局部 GuiState + 手工填充"的范本
（MakeBaselineState()），零 g_state 引用，无需去全局化改造。

- 新增 test/unit-correctness/gui/test_state_reconcile.cpp（3 个 TEST）
- 删除 test/gui/functional/test_gui_state_reconcile.cpp + 注册声明/调用点
- 去掉 test_gui_shared.hpp 依赖：显式 include gui/gui_state.hpp、
  gui/gui_state_reconcile.hpp、<cstdint>，文件内声明 namespace 别名
- 6 处注释里指向 git-ignored 工作制品的任务名改写为机制本身
  （迁移把整份文件变成"新增行"，check_new_refs 会重新扫到它们）

核验：断言 74→74（IM_CHECK 61→EXPECT_TRUE、IM_CHECK_EQ 12→EXPECT_EQ、
IM_CHECK_NE 1→EXPECT_NE）；gui_test 静态用例数 557→554；
gui_unit_test 18→21 全绿。
