+++ test/gui/CMakeLists.txt
@@ -15,7 +15,6 @@ add_executable(gui_test
     functional/test_gui_overlay_labels.cpp
     functional/test_gui_lifecycle.cpp
     functional/test_gui_composite_preview.cpp
-    functional/test_gui_state_reconcile.cpp
     functional/test_gui_sampling_density_stats.cpp
     functional/test_gui_user_defaults.cpp
     functional/test_gui_defaults_panel.cpp
