+++ test/gui/test_gui_shared.hpp
@@ -225,7 +225,6 @@ void RegisterProjectWorldDirTests(ImGuiTestEngine* engine);
 void RegisterHandednessGuardTests(ImGuiTestEngine* engine);
 void RegisterLifecycleTests(ImGuiTestEngine* engine);
 void RegisterCompositePreviewTests(ImGuiTestEngine* engine);
-void RegisterStateReconcileTests(ImGuiTestEngine* engine);
 void RegisterSamplingDensityStatsTests(ImGuiTestEngine* engine);
 void RegisterPreviewAnimationTests(ImGuiTestEngine* engine);
 void RegisterCaptureHarnessTests(ImGuiTestEngine* engine);
